+++ .eslintrc.js
@@ -47,6 +47,9 @@ module.exports = {
         ],
         "array-callback-return": [
           0
-        ]
+        ],
+        "no-use-before-define": [
+          0
+        ],
     }
 };
